requirements.txt
appdirs==1.4.2
cffi==1.9.1
cryptography==1.7.2
dj-database-url==0.4.1
dj-static==0.0.6
Django==1.9.6
django-cors-headers==2.0.2
django-suit==0.2.16
django-toolbelt==0.0.1
djangorestframework==3.3.3
Fabric==1.13.1
gunicorn==19.4.5
idna==2.1
olefile==0.44
packaging==16.8
paramiko==2.1.2
Pillow==4.0.0
psycopg2==2.6.1
pyasn1==0.1.9
pycparser==2.14
pyparsing==2.1.10
six==1.10.0
static3==0.7.0
drfdocs==0.0.11